+++ footprint/src/main/java/com/umc/footprint/src/users/UserController.java
@@ -1,4 +1,4 @@
-package com.umc.footprint.src.users.model;
+package com.umc.footprint.src.users;
 
 
 import org.springframework.beans.factory.annotation.Autowired;
